ai-modules/requirements.txt
// ai-modules/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
requests==2.31.0
pymongo==4.6.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1